+++ src/jtl/Connector/Event/Category/CategoryAfterPullEvent.php
@@ -1,22 +1,24 @@
 <?php
 namespace jtl\Connector\Event\Category;
 
-use \Symfony\Component\EventDispatcher\Event;
-use \jtl\Connector\Model\Category;
+use Symfony\Component\EventDispatcher\Event;
+use jtl\Connector\Model\Category;
+
 
 class CategoryAfterPullEvent extends Event
 {
     const EVENT_NAME = 'category.after.pull';
 
-    protected $category;
+	protected $category;
 
     public function __construct(Category &$category)
     {
-        $this->category = $category;
+		$this->category = $category;
     }
 
     public function getCategory()
     {
         return $this->category;
-    }
+	}
+	
 }