+++ src/jtl/Connector/Event/Category/CategoryBeforeStatisticEvent.php
@@ -1,22 +1,24 @@
 <?php
 namespace jtl\Connector\Event\Category;
 
-use \Symfony\Component\EventDispatcher\Event;
-use \jtl\Connector\Model\Category;
+use Symfony\Component\EventDispatcher\Event;
+use jtl\Connector\Core\Model\QueryFilter;
+
 
 class CategoryBeforeStatisticEvent extends Event
 {
     const EVENT_NAME = 'category.before.statistic';
 
-    protected $category;
+	protected $filter;
 
-    public function __construct(Category &$category)
+    public function __construct(QueryFilter &$filter)
     {
-        $this->category = $category;
+		$this->filter = $filter;
     }
 
-    public function getCategory()
+    public function getFilter()
     {
-        return $this->category;
-    }
+        return $this->filter;
+	}
+	
 }