+++ src/jtl/Connector/Event/Core/CoreConnectorAfterFeaturesEvent.php
@@ -7,15 +7,16 @@ class CoreConnectorAfterFeaturesEvent extends Event
 {
     const EVENT_NAME = 'core.connector.features';
 
-    protected $features;
+	protected $features;
 
     public function __construct(&$features)
     {
-        $this->features = $features;
+		$this->features = $features;
     }
 
     public function getFeatures()
     {
         return $this->features;
-    }
+	}
+	
 }