+++ src/jtl/Connector/Event/DeliveryNote/DeliveryNoteBeforePullEvent.php
@@ -1,22 +1,24 @@
 <?php
 namespace jtl\Connector\Event\DeliveryNote;
 
-use \Symfony\Component\EventDispatcher\Event;
-use \jtl\Connector\Model\DeliveryNote;
+use Symfony\Component\EventDispatcher\Event;
+use jtl\Connector\Core\Model\QueryFilter;
+
 
 class DeliveryNoteBeforePullEvent extends Event
 {
     const EVENT_NAME = 'deliverynote.before.pull';
 
-    protected $deliverynote;
+	protected $filter;
 
-    public function __construct(DeliveryNote &$deliverynote)
+    public function __construct(QueryFilter &$filter)
     {
-        $this->deliverynote = $deliverynote;
+		$this->filter = $filter;
     }
 
-    public function getDeliveryNote()
+    public function getFilter()
     {
-        return $this->deliverynote;
-    }
+        return $this->filter;
+	}
+	
 }