+++ src/jtl/Connector/Event/GlobalData/GlobalDataBeforePushEvent.php
@@ -1,22 +1,24 @@
 <?php
 namespace jtl\Connector\Event\GlobalData;
 
-use \Symfony\Component\EventDispatcher\Event;
-use \jtl\Connector\Model\GlobalData;
+use Symfony\Component\EventDispatcher\Event;
+use jtl\Connector\Model\GlobalData;
+
 
 class GlobalDataBeforePushEvent extends Event
 {
     const EVENT_NAME = 'globaldata.before.push';
 
-    protected $globaldata;
+	protected $globaldata;
 
     public function __construct(GlobalData &$globaldata)
     {
-        $this->globaldata = $globaldata;
+		$this->globaldata = $globaldata;
     }
 
-    public function getGlobalData()
+    public function getGlobaldata()
     {
         return $this->globaldata;
-    }
+	}
+	
 }