+++ src/jtl/Connector/Event/Image/ImageAfterPushEvent.php
@@ -1,22 +1,24 @@
 <?php
 namespace jtl\Connector\Event\Image;
 
-use \Symfony\Component\EventDispatcher\Event;
-use \jtl\Connector\Model\Image;
+use Symfony\Component\EventDispatcher\Event;
+use jtl\Connector\Model\Image;
+
 
 class ImageAfterPushEvent extends Event
 {
     const EVENT_NAME = 'image.after.push';
 
-    protected $image;
+	protected $image;
 
     public function __construct(Image &$image)
     {
-        $this->image = $image;
+		$this->image = $image;
     }
 
     public function getImage()
     {
         return $this->image;
-    }
+	}
+	
 }