+++ src/jtl/Connector/Event/Payment/PaymentAfterStatisticEvent.php
@@ -1,22 +1,24 @@
 <?php
 namespace jtl\Connector\Event\Payment;
 
-use \Symfony\Component\EventDispatcher\Event;
-use \jtl\Connector\Model\Payment;
+use Symfony\Component\EventDispatcher\Event;
+use jtl\Connector\Model\Payment;
+
 
 class PaymentAfterStatisticEvent extends Event
 {
     const EVENT_NAME = 'payment.after.statistic';
 
-    protected $payment;
+	protected $payment;
 
     public function __construct(Payment &$payment)
     {
-        $this->payment = $payment;
+		$this->payment = $payment;
     }
 
     public function getPayment()
     {
         return $this->payment;
-    }
+	}
+	
 }