+++ src/jtl/Connector/Event/ProductPrice/ProductPriceAfterPushEvent.php
@@ -1,22 +1,24 @@
 <?php
 namespace jtl\Connector\Event\ProductPrice;
 
-use \Symfony\Component\EventDispatcher\Event;
-use \jtl\Connector\Model\ProductPrice;
+use Symfony\Component\EventDispatcher\Event;
+use jtl\Connector\Model\ProductPrice;
+
 
 class ProductPriceAfterPushEvent extends Event
 {
     const EVENT_NAME = 'productprice.after.push';
 
-    protected $productprice;
+	protected $productprice;
 
     public function __construct(ProductPrice &$productprice)
     {
-        $this->productprice = $productprice;
+		$this->productprice = $productprice;
     }
 
-    public function getProductPrice()
+    public function getProductprice()
     {
         return $this->productprice;
-    }
+	}
+	
 }