+++ src/jtl/Connector/Event/Specific/SpecificBeforeStatisticEvent.php
@@ -1,22 +1,24 @@
 <?php
 namespace jtl\Connector\Event\Specific;
 
-use \Symfony\Component\EventDispatcher\Event;
-use \jtl\Connector\Model\Specific;
+use Symfony\Component\EventDispatcher\Event;
+use jtl\Connector\Core\Model\QueryFilter;
+
 
 class SpecificBeforeStatisticEvent extends Event
 {
     const EVENT_NAME = 'specific.before.statistic';
 
-    protected $specific;
+	protected $filter;
 
-    public function __construct(Specific &$specific)
+    public function __construct(QueryFilter &$filter)
     {
-        $this->specific = $specific;
+		$this->filter = $filter;
     }
 
-    public function getSpecific()
+    public function getFilter()
     {
-        return $this->specific;
-    }
+        return $this->filter;
+	}
+	
 }