+++ src/jtl/Connector/Event/StatusChange/StatusChangeAfterPushEvent.php
@@ -1,22 +1,24 @@
 <?php
 namespace jtl\Connector\Event\StatusChange;
 
-use \Symfony\Component\EventDispatcher\Event;
-use \jtl\Connector\Model\StatusChange;
+use Symfony\Component\EventDispatcher\Event;
+use jtl\Connector\Model\StatusChange;
+
 
 class StatusChangeAfterPushEvent extends Event
 {
     const EVENT_NAME = 'statuschange.after.push';
 
-    protected $statuschange;
+	protected $statuschange;
 
     public function __construct(StatusChange &$statuschange)
     {
-        $this->statuschange = $statuschange;
+		$this->statuschange = $statuschange;
     }
 
-    public function getStatusChange()
+    public function getStatuschange()
     {
         return $this->statuschange;
-    }
+	}
+	
 }