+++ src/jtl/Connector/Event/StatusChange/StatusChangeBeforePullEvent.php
@@ -1,22 +1,24 @@
 <?php
 namespace jtl\Connector\Event\StatusChange;
 
-use \Symfony\Component\EventDispatcher\Event;
-use \jtl\Connector\Model\StatusChange;
+use Symfony\Component\EventDispatcher\Event;
+use jtl\Connector\Core\Model\QueryFilter;
+
 
 class StatusChangeBeforePullEvent extends Event
 {
     const EVENT_NAME = 'statuschange.before.pull';
 
-    protected $statuschange;
+	protected $filter;
 
-    public function __construct(StatusChange &$statuschange)
+    public function __construct(QueryFilter &$filter)
     {
-        $this->statuschange = $statuschange;
+		$this->filter = $filter;
     }
 
-    public function getStatusChange()
+    public function getFilter()
     {
-        return $this->statuschange;
-    }
+        return $this->filter;
+	}
+	
 }