+++ src/jtl/Connector/Type/Image.php
@@ -22,7 +22,9 @@ class Image extends DataType
             new PropertyInfo('filename', 'string', '', false, false, false),
             new PropertyInfo('relationType', 'ImageRelationType', '', false, false, false),
             new PropertyInfo('remoteUrl', 'string', '', false, false, false),
+            new PropertyInfo('name', 'string', '', false, false, false),
             new PropertyInfo('sort', 'integer', 0, false, false, false),
+            new PropertyInfo('i18ns', '\jtl\Connector\Model\ImageI18n', null, false, false, true),
         );
     }